Extract Whisper::LogSettable
